🧪 test: consolidate and achieve 100% coverage for headless.ts
- Merged `headless-full.test.ts` and `headless-security.test.ts` into `headless.test.ts`.
- Added edge case tests for `execGodotSync` and `execGodotAsync` (missing status, empty stdout/stderr, Unknown error fallback).
- Fixed linting issues by using `unknown` intermediate casts for `any` values in tests.
- Verified 100% branch coverage for `src/godot/headless.ts`.
